add evaluate all formuls and print table

+++ include/Row.h
@@ -2,9 +2,9 @@
 #include "Cell.h"
 
 typedef struct {
-  int row_num;
+  long row_num;
   Cell *cells;
 } Row;
 
-Row *row_create(int row_num, int n_columns);
+Row *row_create(long row_num, int n_columns);
 void row_destroy(Row *row, int n_columns);
